requirements.txt
boto3==1.18.1
botocore==1.21.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.4
django-autoslug==1.9.8
django-crispy-forms==1.12.0
django-social-share==2.2.1
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.15.2
gunicorn==20.1.0
idna==2.8
jmespath==0.10.0
Pillow==6.2.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.5.0
six==1.12.0
urllib3==1.26.6
whitenoise==4.1.2
